Move cell definitions into page.go
Before, I don't think I quite got the abstraction
for cells right because I was trying to detach
them from the concept of pages. I think this is
actually a mistake because cells seem very coupled
with pages - the type of cell is entirely
dependent on the type of page and the entire cell
only contained data on one page.

I'd like to move towards some cell abstraction,
but think that what I'm really interested in is
the entire cell content that I can then decode
into a record (I don't really care whether or not
the cells came from a leaf/interior page, or
whether or not the cell overflowed).
