+++ app/page.go
@@ -14,17 +14,6 @@ const INT_TAB_PAGE = 5
 const LEAF_IDX_PAGE = 10
 const LEAF_TAB_PAGE = 13
 
-type DatabaseHeader struct {
-	HeaderString     [16]byte
-	PageSize         uint16
-	FileWriteVersion uint8
-	FileReadVersion  uint8
-	ReservedSpace    uint8
-	Middle           [38]byte
-	TextEncoding     uint32
-	End              [40]byte
-}
-
 type BTreeHeader struct {
 	PageType                uint8
 	FirstFreeBlock          uint16
@@ -35,9 +24,36 @@ type BTreeHeader struct {
 }
 
 type BTreePage struct {
-	Header           BTreeHeader
-	CellPointerArray []uint16
-	Cells            []Cell
+	Header         BTreeHeader
+	IntIdxCells    []IntIdxCell
+	IntTableCells  []IntTableCell
+	LeafIdxCells   []LeafIdxCell
+	LeafTableCells []LeafTableCell
+}
+
+type LeafTableCell struct {
+	PayloadSize  uint64
+	RowID        uint64
+	Payload      []byte
+	OverflowPage *uint32
+}
+
+type LeafIdxCell struct {
+	LeftPage uint32
+	Key      uint64
+}
+
+type IntTableCell struct {
+	KeyPayloadSize uint64
+	Payload        []byte
+	OverflowPage   *uint32
+}
+
+type IntIdxCell struct {
+	LeftPage       uint32
+	KeyPayloadSize uint64
+	Payload        []byte
+	Key            uint64
 }
 
 // pageNum is zero indexed which may be different to the SQLite standard
@@ -55,6 +71,8 @@ func readBTreePage(databaseFile *os.File, dbHeader DatabaseHeader, pageNum uint3
 	if err := binary.Read(databaseFile, binary.BigEndian, &header); err != nil {
 		fmt.Println("Failed to read integer:", err)
 	}
+
+	page := BTreePage{Header: header}
 
 	// If the page isn't an interior b-tree page, we move back four bytes in the
 	// file because the right most pointer isn't actually included in the header
@@ -74,7 +92,6 @@ func readBTreePage(databaseFile *os.File, dbHeader DatabaseHeader, pageNum uint3
 	}
 
 	// Read the cells
-	var cells []Cell
 	for idx, cellStart := range cellPointerArray {
 		// The cellEnd of the cell is either the start of the next cell, or the
 		// end of the page minus however many bytes are reserved at the end of
@@ -94,12 +111,59 @@ func readBTreePage(databaseFile *os.File, dbHeader DatabaseHeader, pageNum uint3
 			fmt.Println("Failed to read integer:", err)
 		}
 
-		cells = append(cells, readCell(dbHeader, cellBytes, header.PageType))
+		// Between iterations of this loop, we should never hit a different case
+		// here (and if we do, something has gone wrong)
+		switch header.PageType {
+		case LEAF_TAB_PAGE:
+			page.LeafTableCells = append(page.LeafTableCells, readLeafTableCell(dbHeader, cellBytes))
+		case LEAF_IDX_PAGE:
+			page.LeafIdxCells = append(page.LeafIdxCells, readLeafIdxCell(dbHeader, cellBytes))
+		case INT_TAB_PAGE:
+			page.IntTableCells = append(page.IntTableCells, readIntTableCell(dbHeader, cellBytes))
+		case INT_IDX_PAGE:
+			page.IntIdxCells = append(page.IntIdxCells, readIntIdxCell(dbHeader, cellBytes))
+		default:
+			panic("Unknown page type")
+		}
+
 	}
 
-	return BTreePage{
-		Header:           header,
-		CellPointerArray: cellPointerArray,
-		Cells:            cells,
+	return page
+}
+
+func readLeafTableCell(header DatabaseHeader, data []byte) LeafTableCell {
+	var cell LeafTableCell
+
+	var offsetPayloadSize uint16
+	cell.PayloadSize, offsetPayloadSize = decodeVarInt(data[:9])
+
+	var offsetRowID uint16
+	cell.RowID, offsetRowID = decodeVarInt(data[offsetPayloadSize : offsetPayloadSize+9])
+
+	// This doesn't take into account the case where the cell overflows onto the
+	// next page (which means that the last four bytes of the cell are a pointer
+	// to the overflow page) but if we encounter this case we'll end up panicing
+	// just after this
+	cell.Payload = data[offsetPayloadSize+offsetRowID:]
+
+	// There's some somewhat complicated logic to deal with the payload
+	// overflowing onto another page. I'm not going to implement it here and
+	// instead panic if we encounter this case
+	if cell.PayloadSize > uint64(header.PageSize-uint16(header.ReservedSpace)) {
+		panic("Payload overflow")
 	}
+
+	return cell
+}
+
+func readLeafIdxCell(header DatabaseHeader, data []byte) LeafIdxCell {
+	panic("not implemented")
+}
+
+func readIntTableCell(header DatabaseHeader, data []byte) IntTableCell {
+	panic("not implemented")
+}
+
+func readIntIdxCell(header DatabaseHeader, data []byte) IntIdxCell {
+	panic("not implemented")
 }
