+++ app/utils.go
@@ -1,5 +1,16 @@
 package main
 
+type DatabaseHeader struct {
+	HeaderString     [16]byte
+	PageSize         uint16
+	FileWriteVersion uint8
+	FileReadVersion  uint8
+	ReservedSpace    uint8
+	Middle           [38]byte
+	TextEncoding     uint32
+	End              [40]byte
+}
+
 // A varint consists of either zero or more bytes which have the high-order
 // bit set followed by a single byte with the high-order bit clear, or nine
 // bytes, whichever is shorter.
